requirements.txt
streamlit==1.31.0
opencv-python-headless==4.7.0.72
mediapipe==0.10.8
numpy==1.24.3
protobuf>=3.20.3,<4.0.0
pandas
Pillow
matplotlib